% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covSim.R
\name{covSim}
\alias{covSim}
\title{Simulate covariates from a Counter factual model}
\usage{
covSim(CFM, n = 100)
}
\arguments{
\item{CFM}{a Counter-Factual model}

\item{n}{number of observation}
}
\value{
A cumulative Hazard function
}
\description{
A function to simulate covariate values from a Counter Factual Model
}
\details{
This functions extracts the covariates form a counter factual model and
samples from them in order to create a simulated dataset.
}
