% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpmH.R
\name{fpmH}
\alias{fpmH}
\title{Generate a cumulative hazard function from a flexible parametric model}
\usage{
fpmH(CFM, maxTime = 24)
}
\arguments{
\item{CFM}{a Counter-Factual model}

\item{maxTime}{maximum time used in the estimation}
}
\value{
A cumulative Hazard function
}
\description{
A function to estimate the survival function based on parameter estimates.
}
\details{
This functions extracts the baseline (cumulative) hazard parameters from a
counter factual model.  This is used in the simulation of datasets.
}
