% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postEval.R
\name{postEval}
\alias{postEval}
\title{A function to evaluate posterior distribution}
\usage{
postEval(mn, sd, bound = 0, direction = "greater")
}
\arguments{
\item{mn}{distribution mean}

\item{sd}{distribution sd}

\item{bound}{The bound to use in the posterior evaluation (defaults to 0)}
}
\description{
postEval evaluated a distribition based on given mean and standard deviation
}
